Building destroyed flow
